Precompute distillation

+++ trainer/trainer.py
@@ -201,7 +201,7 @@ class Trainer(GenericTrainer):
             target_normal_loss = target[new_classes_indices]
             data_normal_loss = data[new_classes_indices]
 
-            target_distillation_loss = y
+            target_distillation_loss = y.float()
             data_distillation_loss = data
 
             y_onehot = torch.FloatTensor(len(target_normal_loss), self.dataset.classes)
